requirements.txt
Flask==3.0.0
flask-cors==4.0.0
supabase==2.3.0
python-dotenv==1.0.0
werkzeug==3.0.0
PyJWT==2.8.0
razorpay==1.3.0
stripe==7.0.0
reportlab==4.0.7
pandas==2.1.3
openpyxl==3.1.2
twilio==8.10.0
sendgrid==6.11.0
gunicorn==21.2.0
qrcode==7.4.2
Pillow==10.1.0
